refactor(query/memory): extract row-hydration + persistence helpers to hydrate.rs
query/memory/resolve.rs was 1118 lines mixing binding resolution, anchor lookups,
call-path hashing, and low-level memory-record hydration/persistence. Move the
hydration group (memory_row/binding_row/call_path_row/attach_memory_children/
tags_for_memory/ids_to_memories/split_active_stale + replace_tags/upsert_memory_fts/
duplicate_memory_id) into hydrate.rs, re-exported pub(crate) so callers are
unchanged. resolve.rs drops to 938. No behavior change.
